REF: move ShallowMixin to groupby.base
