+++ pandas/core/groupby/base.py
@@ -13,7 +13,22 @@ from pandas.core.base import PandasObject
 OutputKey = collections.namedtuple("OutputKey", ["label", "position"])
 
 
-class GroupByMixin(PandasObject):
+class ShallowMixin(PandasObject):
+    _attributes: List[str] = []
+
+    def _shallow_copy(self, obj, **kwargs):
+        """
+        return a new object with the replacement attributes
+        """
+        if isinstance(obj, self._constructor):
+            obj = obj.obj
+        for attr in self._attributes:
+            if attr not in kwargs:
+                kwargs[attr] = getattr(self, attr)
+        return self._constructor(obj, **kwargs)
+
+
+class GotItemMixin(PandasObject):
     """
     Provide the groupby facilities to the mixed object.
     """
